+++ src/geo/linestring.h
@@ -32,7 +32,7 @@ public:
   LineString(std::initializer_list<atools::geo::Pos> list);
 
   void append(float longitudeX, float latitudeY, float alt = 0.f);
-  void append(double longitudeX, double latitudeY, float alt = 0.f);
+  void append(double longitudeX, double latitudeY, double alt = 0.f);
 
   float lengthMeter() const;
